pyproject.toml
[project]
name = "ost-skg-if-api-tools"
version = "0.1.1"
description = "This project merges multiple OpenAPI specifications extensions into a single core specification for SKG IF."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "pyyaml>=6.0.2",
    "toml>=0.10.2",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "ruff>=0.11.9",
]

[tool.uv.workspace]
members = [
    "test",
]

[tool.uv]
package = true

[project.scripts]
merge = "merge.merge:main"

[tool.setuptools.package-data]
"merge" = ["API/**/*"]